pyproject.toml
[project]
name = "sitefeed"
version = "0.1"
requires-python = ">=3.11"
dependencies = [
    "scrapy~=2.8.0",
    "spidermon[monitoring]~=1.17.1",
    "readability-lxml~=0.8.1",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["sitefeed"]
